+++ src/store/UI_store/UserCategoriesSlice/UserCategoriesInterfaces.ts
@@ -1,0 +1,3 @@
+export interface IUserCategoriesState {
+    currentGroup: string
+}
